Compilation fixes for SERVICE build.
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@59439 91177308-0d34-0410-b5e6-96231b3b80d8

+++ include/mvm/Threads/Thread.h
@@ -92,7 +92,7 @@ public:
 
 #ifdef ISOLATE
   /// IsolateID - The Isolate ID of the thread's VM.
-  uint32 IsolateID;
+  size_t IsolateID;
 #endif
 
   /// MyVM - The VM attached to this Thread.
